requirements.txt
requests==2.26.0
python-dotenv==0.19.2 
boto3==1.20.20
pandas==1.0.3
NumPy==1.21.5